legal-rag-assistant-main/requirements.txt
openai
langchain
langchain_community
faiss-cpu
autogen
pymupdf
langchain_openai
tiktoken
python-dotenv
streamlit